Register repl actions with editor after session started
